img-engine/requirements.txt
aniso8601==9.0.1
attrs==20.3.0
click==7.1.2
itsdangerous==1.1.0
jsonschema==3.2.0
MarkupSafe==1.1.1
pyrsistent==0.17.3
pytz==2021.1
pyzmq==22.0.3
six==1.15.0
zmq==0.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1